Showing the code in text instead of an image in AndroCFG #37
This work adds the possibility to see the decompiled code by Pithus in a raw text (nicely colorized) in the Control Flow Graph section. The code is shown in a large modal.

Unfortunately, due to long lines of code, it is known that the background of the iframe is only limited to the width of the modal, making the line stick out a bit.

Moreover, this code is retrocompatible. In case the sample has been analyzed before this update, Pithus will keep showing the pictures until it has been reanalyzed.

This work has comment on the place holder for the VT key as it is breaking the build in the newer version of Docker Compose.